Move Skifree function to index.html.
